# perf: shapetracker to tuple of views