… tinygrad/ops.py
… tinygrad/shape/shapetracker.py