Conceptual refactor:
* ShapeTrackers into a NamedTuple.
* Reduce inputs of shapetracker methods
* better caching of methods
* Replace Lazycache with small lru_cache (shapetracker is now hashable)

Minor speed gain, ~9,4%

Part of the work towards https://github.com/tinygrad/tinygrad/issues/1209.

Together with other minor changes I should achieve 20%. 

I'll make another MR that combines all changes.

## TODO

- [ ] Tests

## Benchmarks

master
```
codegen         mean runtime:  148.19ms, runs:   164.08,  132.56,  195.60,  142.20,  136.55,  140.03,  140.54,  140.76,  142.42,  147.12
methodcache     mean runtime:  142.41ms, runs:   137.83,  133.41,  190.43,  138.24,  137.90,  132.92,  134.28,  140.02,  140.68,  138.38
profile         mean runtime:  631.34ms, runs:   711.69,  621.92,  613.62,  606.61,  659.79,  610.61,  614.25,  622.88,  629.94,  622.05
```

branch
```
codegen         mean runtime:  121.63ms, runs:   143.90,  115.64,  123.02,  122.06,  118.10,  118.78,  120.01,  119.78,  119.96,  115.06
methodcache     mean runtime:  109.72ms, runs:   108.67,  112.51,  111.18,  110.02,  110.97,  110.00,  111.16,  111.31,  106.20,  105.19
profile         mean runtime:  469.68ms, runs:   487.88,  468.68,  462.18,  469.44,  456.86,  466.85,  465.30,  482.63,  466.75,  470.25
```


master

fastest: 437.51 ms (not counting the first, see https://github.com/tinygrad/tinygrad/issues/1356)

```
ran model in 401.60 ms
sync in 32156.58 ms
 I
ran model in 1498.51 ms
sync in 180.15 ms
'
ran model in 452.35 ms
sync in 335.45 ms
m
ran model in 497.65 ms
sync in 327.55 ms
 a
ran model in 464.49 ms
sync in 333.33 ms
 
ran model in 478.68 ms
sync in 346.87 ms
2
ran model in 458.87 ms
sync in 345.90 ms
0
ran model in 442.85 ms
sync in 345.12 ms
 year
ran model in 437.51 ms
sync in 350.30 ms
 old
ran model in 472.58 ms
sync in 316.63 ms
```

branch

fastest: 396.17 ms (not counting the first, see https://github.com/tinygrad/tinygrad/issues/1356)

```
ran model in 382.85 ms
sync in 31975.68 ms
 I
ran model in 860.98 ms
sync in 306.12 ms
'
ran model in 407.66 ms
sync in 386.02 ms
m
ran model in 418.58 ms
sync in 380.16 ms
 a
ran model in 420.70 ms
sync in 373.58 ms
 
ran model in 434.08 ms
sync in 370.63 ms
2
ran model in 413.64 ms
sync in 373.21 ms
0
ran model in 410.82 ms
sync in 430.54 ms
 year
ran model in 396.17 ms
sync in 395.70 ms
 old
ran model in 420.56 ms
sync in 386.05 ms
```