@tinyb0t 2023-07-27
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py            40              35
tinygrad/codegen/optimizer.py             14              13
tinygrad/lazy.py                          61              54
tinygrad/ops.py                            2               2
tinygrad/shape/shapetracker.py           117             100
tinygrad/tensor.py                         2               2
------------------------------------------------------------
total                                    236             206
------------------------------------------------------------
lines added in the tinygrad folder: 30
```