@geohot 2023-07-31
Going to close, need to see at least 20% speedup or way less lines.